fix(terminal): keep same-frame release-and-press drags extendable
Process primary release events in frame order so a release followed by a
same-frame press hands the still-held pointer to the restarted drag
instead of finishing it, and replay only local-selection anchors after
same-frame scrollback changes so the claimed-press replay can never
reopen Ctrl+click targets. Split terminal input into keyboard, routing,
and frame-events leaf modules to stay inside the maintainability budget.
